fixed things with refactor
